Update FromHob derive macro to utilize zerocopy
This commit updates the FromHob derive macro to rely on zerocopy to do
its conversion from a bytestream to a type. This requires that the
developer also derive implement zerocopy::FromBytes on the type. A
compile time error occurs if that type does not implement FromBytes.

Please note that there is no risk of trait version mismatches like we
have seen in the past, because there the entire usage (implementation
and consumption) is contained within this single file.