GCD: Allow Setting RP on Already Unmapped Pages
Currently, the GCD will debug_assert if RP is attempted to
be applied to a memory region that is already unmapped. It does
this to attempt to catch synchronization errors between the GCD
and the page table. However, this misses that there are legitimate
cases for this pattern, namely when freeing image stacks that have
a guard page set.

With the current GCD behavior, the image code has to remap the guard
page, then call free_pages on the entire range, so that it can be
freed. This is onerous and the current check is not particularly
useful. We should be catching memory that is already freed at a
different layer.

Signed-off-by: Oliver Smith-Denny <osde@microsoft.com>