GCD: Allow Freeing Memory Across GCD Descriptors
Currently, the GCD will fail to free memory if it
crosses descriptors. This is because the underlying
memory block layer has a fundamental assumption it will
work on a single block at a time.

This unnecessarily creates hurdles for consumers and leaks
internal state.

This commit refactors the free_memory_space API to be able
to free memory across descriptors. This is useful for freeing
image memory that has different attributes set across the range,
e.g.

In addition, this unifies the free memory protections across
the free and free_preserving_ownership APIs, the latter was
not properly protected previously.

Signed-off-by: Oliver Smith-Denny <osde@microsoft.com>