AARCH64 GIC v3: add support for booting on a primary core that is not the first entry in the GIC Redistributor. (#1063)
## Description

The prior implementation of the gic_manager module and associated code
assumed that the system boot core would be the first entry in the GIC
redistributor. However, that is not always the case. This PR refactors
the gic_manager module, brings it up to the most recent version of the
`arm-gic` crate, and adds support for booting from a core that is not
the first entry in the GIC redistributor.

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Verified interrupt functionality on AARCH64 hardware platform booting
from a core that isn't the first entry in the GIC redistributor.

## Integration Instructions

N/A