Always enable GICv4 support. (#1074)
## Description

This PR always enables GICv4 support in the arm-gic crate. GicV4 support
is backwards compatible with GicV3, and is necessary for properly
calculating the GIC Redistributor register offsets.

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Booted on hardware with GicV4, observed that GIC redistributor accesses
are no longer incorrect.

## Integration Instructions

N/A - changes are backward compat with GicV3, though the expectation is
that most systems are GicV4.