patina_dxe_core: Prevent setting hob_list from being optimized out
The compiler, on release builds, was optimizing out the call_once call, so future calls to set_hob_list was panicking. This commit changes the debug_assert to an ssert and changes the 	rue return from set_hob_list to be based off of the inner value of hob_list to hint to the compiler that the call_once cannot be optimized away.