patia_adv_logger: Include readme in module docs
This commit does three things:

1. It completely reworks the README to make it more friendly to new
users who will know nothing about the advanced logger. It includes full
setup instructions and guidance on usage.

2. It replaces the top level module documentation for the crate with
this readme. This has two effects. The first is that the example provided
in the readme is now tested and it has the readme and module
documentation always match.

3. It makes the `parser` module available during doc builds so that the
crate documentation on crates.io shows the the `parser` module.