patina_dxe_core: Store loaded image buffer in a Box<[u8]>.
This commit updates the storage device of the `PrivateImageData` struct's
image_buffer to a Box<[u8]> to simplify the lifecycle of a loaded image.

This change allows the Box to manage the lifecycle of the buffer, rather
then needing to store the buffers base page and number of pages purely
for the deallocation process. This enables multiple simplifications of
image.rs