patina_dxe_core: Use MemoryManager Interface for resource_section
Updates image.rs to use the MemoryManager interface for allocating the
pages that hold the HII resource section. This allows us to greatly
reduce the complexity regarding the HII resource section. Deallocation
is now automatic, which allows us to reduce the fields 3 fields in
PrivateImageData related to the HII resource section to only one field,
which are the bytes themselves.